+++ app/javascript/components/training-data/code-tagger/left-pane/LeftPane.tsx
@@ -1,30 +1,17 @@
 import React from 'react'
-import SolutionInfo from './SolutionInfo'
+import CodeInfo from './CodeInfo'
 import { FilePanel } from '@/components/mentoring/session/FilePanel'
 import { CloseButton } from '@/components/mentoring/session/CloseButton'
-import type { CommunitySolution } from '@/components/types'
-import { SolutionTaggerSolution } from '../SolutionTagger.types'
+import { CodeTaggerCode } from '../CodeTagger.types'
 
-export type PanesProps = {
-  solution: CommunitySolution
-}
-
-export function LeftPane({
-  solution,
-}: {
-  solution: SolutionTaggerSolution
-}): JSX.Element {
+export function LeftPane({ code }: { code: CodeTaggerCode }): JSX.Element {
   return (
     <>
       <header className="discussion-header">
         <CloseButton url="#" />
-        <SolutionInfo exercise={solution.exercise} track={solution.track} />
+        <CodeInfo exercise={code.exercise} track={code.track} />
       </header>
-      <FilePanel
-        files={solution.files}
-        language={solution.language}
-        indentSize={2}
-      />
+      <FilePanel files={code.files} language={code.language} indentSize={2} />
     </>
   )
 }
